🐛 버그 수정 및 views 계층 분리
